apps/api/pyproject.toml
[project]
name = "oss-ticketing-api"
version = "0.1.0"
description = "Enterprise OSS ticketing backend (FastAPI + Postgres) with Gmail journal ingestion."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.18.3",
  "bleach>=6.3.0",
  "boto3>=1.42.44",
  "cryptography>=46.0.4",
  "fastapi>=0.128.5",
  "httpx>=0.28.1",
  "opentelemetry-exporter-otlp-proto-http>=1.38.0",
  "opentelemetry-instrumentation-fastapi>=0.59b0",
  "opentelemetry-instrumentation-sqlalchemy>=0.59b0",
  "opentelemetry-sdk>=1.38.0",
  "orjson>=3.11.7",
  "prometheus-client>=0.24.1",
  "psycopg[binary]>=3.3.2",
  "pydantic>=2.12.5",
  "pydantic-settings>=2.12.0",
  "python-multipart>=0.0.22",
  "sqlalchemy>=2.0.46",
  "uvicorn[standard]>=0.40.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.15.0",
]
test = [
  "pytest>=9.0.2",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008", "SIM108"]  # B008: FastAPI Depends pattern, SIM108: ternary preference

[tool.ruff.lint.per-file-ignores]
"alembic/versions/*.py" = ["E501"]  # SQL in migrations can exceed line length
"app/worker/jobs/*.py" = ["E501"]  # SQL in worker jobs can exceed line length

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
